+++ NEWS.md
@@ -1,3 +1,7 @@
-# tickle 0.1.0
+# tickle 0.1.1  2022-06-03
+
+* Add `idle` callback to `tic_window()`
+
+# tickle 0.1.0  2022-06-03
 
 * Initial release